feat(lynx): merge @rozenite/lynx-dev in and inject the device runtime
Setting Rozenite up for Lynx cost two installs and a hand-written import in
the app's entry point. That import had to be `__DEV__`-guarded or the
dispatcher shipped to production, and the README offered the unguarded form
as an equally valid alternative.

`@rozenite/lynx` now carries both halves: `.` stays the device runtime,
`./rspeedy` is the plugin formerly published as `@rozenite/lynx-dev`. The
plugin injects the runtime itself via `source.preEntry`, from inside a
`setup` that `apply: 'serve'` already keeps out of `rspeedy build` — so the
production leak is structurally impossible rather than documented against.

`@rozenite/lynx-dev` stays as a deprecated re-export shim.

Closes #488
