+++ apps/playground-lynx/lynx.config.ts
@@ -3,7 +3,7 @@ import { defineConfig } from '@lynx-js/rspeedy';
 import { pluginQRCode } from '@lynx-js/qrcode-rsbuild-plugin';
 import { pluginReactLynx } from '@lynx-js/react-rsbuild-plugin';
 import { pluginTypeCheck } from '@rsbuild/plugin-type-check';
-import { rozeniteLynxPlugin } from '@rozenite/lynx-dev';
+import { rozeniteLynxPlugin } from '@rozenite/lynx/rspeedy';
 
 export default defineConfig({
   plugins: [
