+++ apps/playground-lynx/package.json
@@ -14,7 +14,6 @@
     "@lynx-js/react": "^0.125.0",
     "@rozenite/controls-plugin": "workspace:*",
     "@rozenite/feature-flags-plugin": "workspace:*",
-    "@rozenite/lynx": "workspace:*",
     "@rozenite/rhf-plugin": "workspace:*",
     "@rozenite/tanstack-query-plugin": "workspace:*",
     "@tanstack/react-query": "^5.81.5",
@@ -26,7 +25,7 @@
     "@lynx-js/react-rsbuild-plugin": "^0.19.1",
     "@lynx-js/rspeedy": "^0.16.5",
     "@lynx-js/types": "4.1.0",
-    "@rozenite/lynx-dev": "workspace:*",
+    "@rozenite/lynx": "workspace:*",
     "@rsbuild/plugin-type-check": "1.6.0",
     "@types/react": "^19.2.18",
     "eslint": "^9.25.0",
