+++ packages/lynx-dev/tsconfig.json
@@ -4,7 +4,7 @@
   "include": [],
   "references": [
     {
-      "path": "../middleware"
+      "path": "../lynx"
     },
     {
       "path": "./tsconfig.lib.json"
